+++ htdocs/CreateRenewalsSalesOrders.php
@@ -1,13 +1,13 @@
 <?php
 /**
-* Renewals sales order creation front end controller
-* CNC Ltd
-*
-* Generates renewals
-*
-* @access public
-* @authors Karim Ahmed - Sweet Code Limited
-*/
+ * Renewals sales order creation front end controller
+ * CNC Ltd
+ *
+ * Generates renewals
+ *
+ * @access public
+ * @authors Karim Ahmed - Sweet Code Limited
+ */
 require_once("config.inc.php");
 GLOBAL $cfg;
 require_once($cfg['path_bu'] . '/BURenBroadband.inc.php');
@@ -21,12 +21,14 @@ $buRenQuotation = new BURenQuotation($this);
 $buRenDomain = new BURenDomain($this);
 $buRenHosting = new BURenHosting($this);
 
-$buRenBroadband->emailRenewalsSalesOrdersDue();
-$buRenContract->emailRenewalsSalesOrdersDue();
-$buRenHosting->emailRenewalsSalesOrdersDue();
-$buRenQuotation->emailRenewalsQuotationsDue();
-$buRenQuotation->emailRecentlyGeneratedQuotes();
-$buRenDomain->emailRenewalsSalesOrdersDue();
+$toEmail = "CreateRenewalSalesOrders@cnc-ltd.co.uk";
+
+$buRenBroadband->emailRenewalsSalesOrdersDue($toEmail);
+$buRenContract->emailRenewalsSalesOrdersDue($toEmail);
+$buRenHosting->emailRenewalsSalesOrdersDue($toEmail);
+$buRenQuotation->emailRenewalsQuotationsDue($toEmail);
+$buRenQuotation->emailRecentlyGeneratedQuotes($toEmail);
+$buRenDomain->emailRenewalsSalesOrdersDue($toEmail);
 
 $buRenBroadband->createRenewalsSalesOrders();
 $buRenContract->createRenewalsSalesOrders();